chore(search-page): move filters to separate file
